+++ src/main/java/com/softbistro/survey/question/controller/QuestionSectionController.java
@@ -1,12 +1,15 @@
 package com.softbistro.survey.question.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.softbistro.survey.client.auth.service.AuthorizationService;
 import com.softbistro.survey.question.components.entity.QuestionSection;
 import com.softbistro.survey.question.service.QuestionSectionService;
 import com.softbistro.survey.response.Response;
@@ -20,6 +23,11 @@ import com.softbistro.survey.response.Response;
 @RestController
 @RequestMapping("/rest/survey/v1/questionSection")
 public class QuestionSectionController {
+	
+	private static final String UNAUTHORIZED_CLIENT = "Unauthorized client";
+	
+	@Autowired
+	private AuthorizationService authorizationService;
 
 	@Autowired
 	private QuestionSectionService questionSectionService;
@@ -31,7 +39,13 @@ public class QuestionSectionController {
 	 * @return Response
 	 */
 	@RequestMapping(method = RequestMethod.POST)
-	public Response setQuestionSection(@RequestBody QuestionSection questionSection) {
+	public Response setQuestionSection(@RequestBody QuestionSection questionSection, @RequestHeader String token) {
+		
+		if (!authorizationService.checkAccess(token)) {
+
+			return new Response(null, HttpStatus.OK, UNAUTHORIZED_CLIENT);
+		}
+		
 		return questionSectionService.setQuestionSection(questionSection);
 	}
 
@@ -43,7 +57,13 @@ public class QuestionSectionController {
 	 */
 	@RequestMapping(value = "/{id}", method = RequestMethod.PUT)
 	public Response updateQuestionSection(@RequestBody QuestionSection questionSection,
-			@PathVariable("id") Integer questionSectionId) {
+			@PathVariable("id") Integer questionSectionId, @RequestHeader String token) {
+		
+		if (!authorizationService.checkAccess(token)) {
+
+			return new Response(null, HttpStatus.OK, UNAUTHORIZED_CLIENT);
+		}
+		
 		return questionSectionService.updateQuestionSection(questionSection, questionSectionId);
 	}
 
@@ -54,7 +74,13 @@ public class QuestionSectionController {
 	 * @return Response
 	 */
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public Response deleteQuestionSection(@PathVariable("id") Integer questionSectionId) {
+	public Response deleteQuestionSection(@PathVariable("id") Integer questionSectionId, @RequestHeader String token) {
+		
+		if (!authorizationService.checkAccess(token)) {
+
+			return new Response(null, HttpStatus.OK, UNAUTHORIZED_CLIENT);
+		}
+		
 		return questionSectionService.deleteQuestionSection(questionSectionId);
 	}
 
@@ -65,7 +91,13 @@ public class QuestionSectionController {
 	 * @return Response
 	 */
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public Response getQuestionSectionById(@PathVariable Integer id) {
+	public Response getQuestionSectionById(@PathVariable Integer id, @RequestHeader String token) {
+		
+		if (!authorizationService.checkAccess(token)) {
+
+			return new Response(null, HttpStatus.OK, UNAUTHORIZED_CLIENT);
+		}
+		
 		return questionSectionService.getQuestionSectionById(id);
 	}
 
@@ -76,7 +108,13 @@ public class QuestionSectionController {
 	 * @return Response
 	 */
 	@RequestMapping(value = "/survey/{id}", method = RequestMethod.GET)
-	public Response getQuestionSectionBySurveyId(@PathVariable Integer id) {
+	public Response getQuestionSectionBySurveyId(@PathVariable Integer id, @RequestHeader String token) {
+		
+		if (!authorizationService.checkAccess(token)) {
+
+			return new Response(null, HttpStatus.OK, UNAUTHORIZED_CLIENT);
+		}
+		
 		return questionSectionService.getQuestionSectionBySurveyId(id);
 	}
 
@@ -87,7 +125,13 @@ public class QuestionSectionController {
 	 * @return Response
 	 */
 	@RequestMapping(value = "/name/{name}", method = RequestMethod.GET)
-	public Response getQuestionSectionById(@PathVariable String name) {
+	public Response getQuestionSectionById(@PathVariable String name, @RequestHeader String token) {
+		
+		if (!authorizationService.checkAccess(token)) {
+
+			return new Response(null, HttpStatus.OK, UNAUTHORIZED_CLIENT);
+		}
+		
 		return questionSectionService.getQuestionSectionByName(name);
 	}
 }
